RAG-feat-Post/requirements.txt
pdfplumber
chromadb
sentence_transformers
python-docx
beautifulsoup4
fastapi
uvicorn